pyproject.toml
[tool.poetry]
name = "nixos_security_tracker"
version = "0.1.0"
description = ""
authors = ["Andreas Rammhold <andreas@rammhold.de>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
django-bootstrap4 = "^2.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-django = "^4.1.0"
django-stubs = "^1.7.0"
factory-boy = "^3.1.0"

[tool.poetry.scripts]
# value needs a colon (:)
# nixos-security-tracker = "nixos_security_tracker.wsgi"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# [tool.isort]
# profile = "black"

[tool.flake8]
max-line-length = 120
